service_check: refactor
